# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Persian CRM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /equipment-categories:
    get:
      operationId: listEquipmentCategories
      tags: [equipment]
      summary: List all equipment categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EquipmentCategory"
    post:
      operationId: createEquipmentCategory
      tags: [equipment]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentCategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquipmentCategory"
  /equipment-categories/{id}:
    patch:
      operationId: updateEquipmentCategory
      tags: [equipment]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentCategoryInput"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquipmentCategory"
    delete:
      operationId: deleteEquipmentCategory
      tags: [equipment]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /equipment:
    get:
      operationId: listEquipment
      tags: [equipment]
      summary: List all equipment with advanced search
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of equipment
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EquipmentWithPrices"
    post:
      operationId: createEquipment
      tags: [equipment]
      summary: Create equipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentInput"
      responses:
        "201":
          description: Created equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquipmentWithPrices"
  /equipment/{id}:
    get:
      operationId: getEquipment
      tags: [equipment]
      summary: Get equipment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquipmentWithPrices"
        "404":
          description: Not found
    patch:
      operationId: updateEquipment
      tags: [equipment]
      summary: Update equipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentInput"
      responses:
        "200":
          description: Updated equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquipmentWithPrices"
    delete:
      operationId: deleteEquipment
      tags: [equipment]
      summary: Delete equipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /equipment/{id}/suppliers:
    get:
      operationId: listEquipmentSuppliers
      tags: [equipment]
      summary: List suppliers for equipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of equipment-supplier links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EquipmentSupplier"
    post:
      operationId: addEquipmentSupplier
      tags: [equipment]
      summary: Add supplier to equipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentSupplierInput"
      responses:
        "201":
          description: Created link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquipmentSupplier"
  /equipment/{id}/suppliers/{linkId}:
    patch:
      operationId: updateEquipmentSupplier
      tags: [equipment]
      summary: Update equipment-supplier link
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: linkId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentSupplierInput"
      responses:
        "200":
          description: Updated link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquipmentSupplier"
    delete:
      operationId: deleteEquipmentSupplier
      tags: [equipment]
      summary: Delete equipment-supplier link
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: linkId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    EquipmentCategory:
      type: object
      required: [id, name, created_at]
      properties:
        id:
          type: integer
        name:
          type: string
        created_at:
          type: string
    EquipmentCategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
    EquipmentSpec:
      type: object
      required: [key, value]
      properties:
        key:
          type: string
        value:
          type: string
    Equipment:
      type: object
      required: [id, name, specs, created_at]
      properties:
        id:
          type: integer
        name:
          type: string
        category_id:
          type: ["integer", "null"]
        category_name:
          type: ["string", "null"]
        specs:
          type: array
          items:
            $ref: "#/components/schemas/EquipmentSpec"
        description:
          type: ["string", "null"]
        created_at:
          type: string
    EquipmentWithPrices:
      allOf:
        - $ref: "#/components/schemas/Equipment"
        - type: object
          properties:
            min_purchase_price:
              type: ["number", "null"]
            min_sell_price:
              type: ["number", "null"]
            default_supplier_name:
              type: ["string", "null"]
            default_brand:
              type: ["string", "null"]
    EquipmentInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        category_id:
          type: ["integer", "null"]
        specs:
          type: array
          items:
            $ref: "#/components/schemas/EquipmentSpec"
        description:
          type: string
    EquipmentSupplier:
      type: object
      required: [id, equipment_id, supplier_id, purchase_price, sell_price, is_default]
      properties:
        id:
          type: integer
        equipment_id:
          type: integer
        supplier_id:
          type: integer
        supplier_name:
          type: ["string", "null"]
        brand:
          type: ["string", "null"]
        purchase_price:
          type: number
        sell_price:
          type: number
        supplier_code:
          type: ["string", "null"]
        is_default:
          type: boolean
        notes:
          type: ["string", "null"]
    EquipmentSupplierInput:
      type: object
      required: [supplier_id, purchase_price, sell_price]
      properties:
        supplier_id:
          type: integer
        brand:
          type: string
        purchase_price:
          type: number
        sell_price:
          type: number
        supplier_code:
          type: string
        is_default:
          type: boolean
        notes:
          type: string
...