# lib/db/src/schema/equipment_categories.ts
import { pgTable, serial, text, timestamp } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const equipmentCategoriesTable = pgTable("equipment_categories", {
  id: serial("id").primaryKey(),
  name: text("name").notNull(),
  created_at: timestamp("created_at").notNull().defaultNow(),
});

export const insertEquipmentCategorySchema = createInsertSchema(equipmentCategoriesTable).omit({ id: true, created_at: true });
export type InsertEquipmentCategory = z.infer<typeof insertEquipmentCategorySchema>;
export type EquipmentCategory = typeof equipmentCategoriesTable.$inferSelect;